requirement.txt
# pip freeze > requirement.txt              将已安装的库写入文件
# pip freeze                                显示已安装库的版本
# pip install -r requirement.txt            批量安装文件中的所有库
# pip uninstall -r requirement.txt          批量卸载文件中的所有库
Flask==1.1.1
mysql-connector-python==8.0.17
openpyxl==2.6.2
PyMySQL==0.9.3
PyYAML==5.1.2
requests==2.22.0
selenium==3.141.0
xlrd==1.2.0
xlwt==1.3.0

robotframework==3.1.2
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==3.3.1